Started adding sprites. Maybe they work.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@226 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/GSprmtvs.h
@@ -37,7 +37,7 @@ const int pr_quadlist       =8;     //GL_QUADS
 const int pr_quadstrip      =9;     //GL_QUAD_STRIP
 const int pr_polygon       =10;     //GL_POLYGON
 
-int draw_primitive_begin(double kind);
+int draw_primitive_begin(int kind);
 int draw_vertex(double x, double y);
 int draw_vertex_color(float x, float y, int color, float alpha);
 int draw_primitive_end();
